requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
deep-translator==1.11.4
idna==3.7
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
mdurl==0.1.2
mysql-connector-python==9.0.0
Pygments==2.18.0
pyproject-toml==0.0.10
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.19.0
shellingham==1.5.4
soupsieve==2.5
toml==0.10.2
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.1
